4.15-4.22 solutions

+++ part4/bloglist-backend/controllers/blogs.js
@@ -2,25 +2,46 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', {
+        username: 1,
+        name: 1,
+        id: 1,
+    })
     response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
+    const token = request.token
+    const user = request.user
+    if (!token || !user.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
 
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
         likes: body.likes,
+        user: user._id,
     })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const blog = await Blog.findById(request.params.id)
+    const user = request.user
+    if (blog.user.toString() !== user.id) {
+        return response
+            .status(403)
+            .json({ error: 'Only own created blogs can be deleted' })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
